MULE-17507: Remove unused legacy processing strategies (#8251)
