+++ core-tests/src/test/java/org/mule/runtime/core/internal/construct/DefaultFlowTestCase.java
@@ -38,8 +38,6 @@ import org.mule.runtime.api.lifecycle.Disposable;
 import org.mule.runtime.api.lifecycle.LifecycleException;
 import org.mule.runtime.api.lifecycle.Startable;
 import org.mule.runtime.api.lifecycle.Stoppable;
-import org.mule.runtime.api.scheduler.Scheduler;
-import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.construct.Flow;
 import org.mule.runtime.core.api.construct.FlowConstruct;
 import org.mule.runtime.core.api.event.CoreEvent;
@@ -50,7 +48,6 @@ import org.mule.runtime.core.api.processor.strategy.ProcessingStrategy;
 import org.mule.runtime.core.api.source.MessageSource;
 import org.mule.runtime.core.internal.construct.DefaultFlowBuilder.DefaultFlow;
 import org.mule.runtime.core.internal.processor.strategy.BlockingProcessingStrategyFactory;
-import org.mule.runtime.core.internal.processor.strategy.WorkQueueStreamProcessingStrategyFactory;
 import org.mule.runtime.core.internal.transformer.simple.StringAppendTransformer;
 import org.mule.tck.SensingNullMessageProcessor;
 import org.mule.tck.core.lifecycle.LifecycleTrackerProcessor;
@@ -58,9 +55,7 @@ import org.mule.tck.core.lifecycle.LifecycleTrackerProcessor;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.RejectedExecutionException;
 import java.util.function.BiFunction;
-import java.util.function.Supplier;
 
 import org.junit.After;
 import org.junit.Ignore;
@@ -338,39 +333,6 @@ public class DefaultFlowTestCase extends AbstractFlowConstructTestCase {
     } catch (LifecycleException e) {
       assertThat(e.getCause(), instanceOf(MuleException.class));
       assertThat(e.getCause().getCause(), sameInstance(startException));
-    }
-  }
-
-
-  @Test
-  public void workQueueSchedulerRejectsDoesntStartFlow() throws Exception {
-    Processor processor = mock(Processor.class, withSettings().extraInterfaces(Startable.class, Stoppable.class));
-
-    final Scheduler rejectingScheduler = mock(Scheduler.class);
-    doThrow(new RejectedExecutionException("rejected by test")).when(rejectingScheduler).execute(any());
-
-    final ProcessingStrategy processingStrategy = new WorkQueueStreamProcessingStrategyFactory() {
-
-      @Override
-      protected Supplier<Scheduler> getRingBufferSchedulerSupplier(MuleContext muleContext, String schedulersNamePrefix) {
-        return () -> rejectingScheduler;
-      }
-    }.create(muleContext, FLOW_NAME);
-
-    flow = (DefaultFlow) Flow.builder(FLOW_NAME, muleContext)
-        .source(directInboundMessageSource)
-        .processors(singletonList(processor))
-        .processingStrategyFactory((muleContext, s) -> processingStrategy)
-        .build();
-
-    flow.initialise();
-
-    expectedException.expectMessage("No subscriptions active for processor.");
-
-    try {
-      flow.start();
-    } finally {
-      verify(rejectingScheduler).shutdownNow();
     }
   }
 }
