+++ core/src/main/java/org/mule/runtime/core/internal/processor/strategy/ProactorStreamEmitterProcessingStrategyFactory.java
@@ -13,7 +13,6 @@ import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.BLOCKING;
 import static org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType.CPU_INTENSIVE;
 import static org.mule.runtime.core.internal.context.thread.notification.ThreadNotificationLogger.THREAD_NOTIFICATION_LOGGER_CONTEXT_KEY;
-import static org.slf4j.LoggerFactory.getLogger;
 import static reactor.core.publisher.Flux.from;
 import static reactor.core.publisher.FluxSink.OverflowStrategy.BUFFER;
 import static reactor.core.publisher.Mono.subscriberContext;
@@ -30,6 +29,7 @@ import org.mule.runtime.core.api.event.CoreEvent;
 import org.mule.runtime.core.api.processor.ReactiveProcessor;
 import org.mule.runtime.core.api.processor.ReactiveProcessor.ProcessingType;
 import org.mule.runtime.core.api.processor.Sink;
+import org.mule.runtime.core.api.processor.strategy.AsyncProcessingStrategyFactory;
 import org.mule.runtime.core.api.processor.strategy.ProcessingStrategy;
 import org.mule.runtime.core.internal.context.thread.notification.ThreadLoggingExecutorServiceDecorator;
 
@@ -40,16 +40,14 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.IntUnaryOperator;
 import java.util.function.Supplier;
 
-import org.slf4j.Logger;
-
 import reactor.core.publisher.EmitterProcessor;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 /**
- * Creates {@link ReactorProcessingStrategyFactory.ReactorProcessingStrategy} instance that implements the proactor pattern by
+ * Creates {@link AsyncProcessingStrategyFactory} instance that implements the proactor pattern by
  * de-multiplexing incoming events onto a multiple emitter using the {@link SchedulerService#cpuLightScheduler()} to process these
- * events from each emitter. In contrast to the {@link ReactorStreamProcessingStrategy} the proactor pattern treats
+ * events from each emitter. In contrast to the {@link AbstractStreamProcessingStrategyFactory} the proactor pattern treats
  * {@link ProcessingType#CPU_INTENSIVE} and {@link ProcessingType#BLOCKING} processors differently and schedules there execution
  * on dedicated {@link SchedulerService#cpuIntensiveScheduler()} and {@link SchedulerService#ioScheduler()} ()} schedulers.
  * <p/>
@@ -57,7 +55,7 @@ import reactor.core.publisher.Mono;
  *
  * @since 4.2.0
  */
-public class ProactorStreamEmitterProcessingStrategyFactory extends ReactorStreamProcessingStrategyFactory {
+public class ProactorStreamEmitterProcessingStrategyFactory extends AbstractStreamProcessingStrategyFactory {
 
   @Override
   public ProcessingStrategy create(MuleContext muleContext, String schedulersNamePrefix) {
@@ -77,18 +75,11 @@ public class ProactorStreamEmitterProcessingStrategyFactory extends ReactorStrea
   }
 
   @Override
-  protected int resolveParallelism() {
-    return Integer.max(CORES, getMaxConcurrency());
-  }
-
-  @Override
   public Class<? extends ProcessingStrategy> getProcessingStrategyType() {
     return ProactorStreamEmitterProcessingStrategy.class;
   }
 
   static class ProactorStreamEmitterProcessingStrategy extends ProactorStreamProcessingStrategy {
-
-    private static Logger LOGGER = getLogger(ProactorStreamEmitterProcessingStrategy.class);
 
     private final int bufferSize;
     private final boolean isThreadLoggingEnabled;
@@ -144,7 +135,7 @@ public class ProactorStreamEmitterProcessingStrategyFactory extends ReactorStrea
                                      () -> awaitSubscribersCompletion(flowConstruct, shutdownTimeout, completionLatch,
                                                                       currentTimeMillis()),
                                      createOnEventConsumer(), getBufferQueueSize());
-        sinks.add(new ProactorSinkWrapper<>(sink));
+        sinks.add(sink);
       }
 
       return new RoundRobinReactorSink<>(sinks);
